refactor: extract store query sql helpers
